Move `MapObject` to `libOPHD`
