+++ appOPHD/UI/DetailMap.cpp
@@ -4,7 +4,8 @@
 #include "../Map/Tile.h"
 #include "../Map/TileMap.h"
 #include "../Map/MapView.h"
-#include "../MapObjects/MapObject.h"
+
+#include <libOPHD/MapObjects/MapObject.h>
 
 #include <NAS2D/Renderer/Color.h>
 #include <NAS2D/Renderer/Renderer.h>
