+++ libOPHD/MapObjects/MapObject.cpp
