+++ libOPHD/MapObjects/MapObject.h
