+++ libOPHD/libOPHD.vcxproj.filters
@@ -45,6 +45,9 @@
     <ClCompile Include="Map\MapCoordinate.cpp">
       <Filter>Source Files\Map</Filter>
     </ClCompile>
+    <ClCompile Include="MapObjects\MapObject.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
     <ClCompile Include="MapObjects\OreDeposit.cpp">
       <Filter>Source Files\MapObjects</Filter>
     </ClCompile>
@@ -115,6 +118,9 @@
     </ClInclude>
     <ClInclude Include="Map\MapOffset.h">
       <Filter>Header Files\Map</Filter>
+    </ClInclude>
+    <ClInclude Include="MapObjects\MapObject.h">
+      <Filter>Header Files\MapObjects</Filter>
     </ClInclude>
     <ClInclude Include="MapObjects\StructureType.h">
       <Filter>Header Files\Map</Filter>
